anderson_agent/presentation_chatbot/pyproject.toml
[tool.poetry]
name = "presentation-chatbot"
version = "0.1.0"
description = "Presentation chatbot using Agent Development Kit and Gemini 2.5 Pro"
authors = ["Your Name <your.email@example.com>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.8.1"
tabulate = "^0.9.0"
cloudpickle = "^3.1.1"
pylint = "^3.3.6"
google-cloud-aiplatform = { extras = [
    "adk",
    "agent_engines",
], version = "^1.91.0" }
google-adk = "^1.0.0"
jsonschema = "^4.23.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
scikit-learn = "^1.6.1"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.3"
flake8-pyproject = "^1.2.3"
pylint = "^3.3.6"
pyink = "^24.10.1"
jsonschema = "^4.23.0"
google-cloud-aiplatform = { extras = ["evaluation"], version = "^1.91.0" }

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"
python-dotenv = "^1.0.1"

[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
testpaths = ["tests/", "eval/"]
markers = ["unit"]
log_level = "ERROR"
log_cli = false
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = ["ignore::UserWarning"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.mason"

[tool.pyink]
line-length = 80
pyink-indentation = 4
pyink-use-majority-quotes = true
